Split DataServlet into CommentList and CommentPost servlets
