+++ portfolio/src/main/webapp/index.html
@@ -133,7 +133,7 @@
     <h3>Leave a comment!</h3>
 
     <div>
-      <form method="POST" action="/commentsData">
+      <form method="POST" action="/commentPost">
         <textarea placeholder="Leave a comment" name="comment" required="required"></textarea>
         <button type="submit">Submit</button>
       </form>
